requirements.txt
apache-airflow==2.8.1
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
pandas==2.1.4
pyarrow==14.0.2
numpy>=1.24.0
dbt-core==1.7.4
dbt-bigquery==1.7.4
streamlit==1.30.0
plotly==5.18.0
python-dotenv==1.0.0
structlog==24.1.0
faker==22.0.0
pytest==7.4.4
great-expectations==0.18.8
